Move utils to dedicated directory
One level up, updated documentation as well.
