+++ content/reducers/blogReducer.js
@@ -2,7 +2,7 @@ import {
   rankingEngine as rankSnippet,
   searchIndexingEngine as tokenizeSnippet
 } from 'engines';
-import { convertToSeoSlug, uniqueElements } from 'functions/utils';
+import { convertToSeoSlug, uniqueElements } from 'utils';
 import { determineExpertiseFromTags } from 'functions/transformers';
 // TODO: Consider parsing this via a new parser or similar
 // The argument against is that it's a single case and might not extend to other repos in the future
