+++ src/components/atoms/button/codepenButton/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'typedefs/proptypes';
-import { combineClassNames } from 'functions/utils';
+import { combineClassNames } from 'utils';
 import literals from 'lang/en/client/common';
 
 /* eslint-disable camelcase */
