+++ src/components/molecules/search/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'typedefs/proptypes';
 import { connect } from 'react-redux';
-import { combineClassNames, getURLParameters, throttle, getBaseURL, getRootURL } from 'functions/utils';
+import { combineClassNames, getURLParameters, throttle, getBaseURL, getRootURL } from 'utils';
 import { pushNewQuery, searchByKeyphrase } from 'state/search';
 import { useFetchSearchIndex } from 'components/hooks';
 import { AnchorButton } from 'components/atoms/button';
