+++ src/styles/_icon_colors.scss
@@ -3,38 +3,44 @@
       background: #1f253d;
       color: #edf0fc;
     }
-
+    
     .card-icon.icon-js {
       background: #f6d854;
       color: #392f31;
     }
-
+    
+    .card-icon.icon-js {
+      background: #f6d854;
+      color: #392f31;
+    }
+    
     .card-icon.icon-csharp {
       background: #672179;
       color: #ffffff;
     }
-
+    
     .card-icon.icon-css {
       background: #3f4de4;
       color: #ffffff;
     }
-
+    
     .card-icon.icon-dart {
       background: #1b2634;
       color: #ffffff;
     }
-
+    
     .card-icon.icon-php {
       background: #8b9bd6;
       color: #2a2843;
     }
-
+    
     .card-icon.icon-python {
       background: #3c77a9;
       color: #ffffff;
     }
-
+    
     .card-icon.icon-react {
       background: #282c34;
       color: #61dafb;
     }
+
