+++ src/utils/string.test.js
